--- src/pages/privacy-policy.tsx
@@ -1,226 +1,0 @@
-import Head from "next/head";
-import Link from "next/link";
-import Container from "../components/container";
-import SideBar from "../components/side-bar";
-import { Post } from "../interfaces/post";
-import { getAllPosts } from "../lib/api";
-import { BASE_URL } from "../lib/constants";
-
-type Props = {
-  allPosts: Post[];
-};
-
-export default function PrivacyPolicy({ allPosts }: Props) {
-  const titleText = `Privacy Policy | Best Seen on Foot`;
-
-  return (
-    <>
-      <Head>
-        <title>{titleText}</title>
-        <link
-          rel="canonical"
-          href={`${BASE_URL}/privacy-policy`}
-          key="canonical"
-        />
-      </Head>
-      <Container>
-        <section className="container prose mx-auto mb-32 max-w-none md:prose-lg lg:prose-xl lg:col-span-2">
-          <h1>Privacy Policy</h1>
-          <article>
-            <h2>Who we are</h2>
-
-            <p>
-              Our website address is <Link href="/">{BASE_URL}</Link>
-            </p>
-
-            <h2>What personal data we collect and why we collect it</h2>
-
-            {/* <h3>Comments</h3>
-
-          <p>
-            When visitors leave comments on the site we collect the data shown
-            in the comments form, and also the visitor&apos;s IP address and
-            browser user agent string to help spam detection.
-          </p>
-
-          <p>
-            An anonymized string created from your email address (also called
-            a hash) may be provided to the Gravatar service to see if you are
-            using it. The Gravatar service privacy policy is available here:{" "}
-            <Link
-              href="https://automattic.com/privacy/"
-              target="_blank"
-              rel="noreferrer noopener"
-            >
-              https://automattic.com/privacy/
-            </Link>
-            . After approval of your comment, your profile picture is visible
-            to the public in the context of your comment.
-          </p> */}
-
-            {/* <h3>Media</h3>
-
-          <p>
-            If you upload images to the website, you should avoid uploading
-            images with embedded location data (EXIF GPS) included. Visitors
-            to the website can download and extract any location data from
-            images on the website.
-          </p> */}
-
-            <h3>Contact forms</h3>
-
-            <p>
-              We keep contact form submissions for as long as needed for
-              customer service purposes, but we do not use the information
-              submitted through them for marketing purposes
-            </p>
-
-            <h3>Cookies</h3>
-
-            {/* <p>
-            If you leave a comment on our site you may opt-in to saving your
-            name, email address and website in cookies. These are for your
-            convenience so that you do not have to fill in your details again
-            when you leave another comment. These cookies will last for one
-            year.
-          </p>
-
-          <p>
-            If you have an account and you log in to this site, we will set a
-            temporary cookie to determine if your browser accepts cookies.
-            This cookie contains no personal data and is discarded when you
-            close your browser.
-          </p>
-
-          <p>
-            When you log in, we will also set up several cookies to save your
-            login information and your screen display choices. Login cookies
-            last for two days, and screen options cookies last for a year. If
-            you select &#8220;Remember Me&#8221;, your login will persist for
-            two weeks. If you log out of your account, the login cookies will
-            be removed.
-          </p>
-
-          <p>
-            If you edit or publish an article, an additional cookie will be
-            saved in your browser. This cookie includes no personal data and
-            simply indicates the post ID of the article you just edited. It
-            expires after 1 day.
-          </p> */}
-
-            <p>
-              We have a few other tracking cookies related to social media sites
-              like Facebook, Instagram, and Twitter. There are also some site
-              statistics, advertising, and tracking cookies being run by Google.
-            </p>
-
-            <p>
-              You can always use browser tools to delete and block cookies.{" "}
-              <Link
-                rel="noopener noreferrer"
-                href="http://optout.networkadvertising.org/"
-                target="_blank"
-              >
-                NAI Opt Out
-              </Link>{" "}
-              or{" "}
-              <Link
-                rel="noopener noreferrer"
-                href="http://optout.aboutads.info/"
-                target="_blank"
-              >
-                DDA Opt Out
-              </Link>{" "}
-              are also good ways to clean up a lot of tracking and advertising
-              cookies on your browser.
-            </p>
-
-            <h3>Embedded content from other websites</h3>
-
-            <p>
-              Articles on this site may include embedded content (e.g. videos,
-              images, articles, etc.). Embedded content from other websites
-              behaves in the exact same way as if the visitor has visited the
-              other website.
-            </p>
-
-            <p>
-              These websites may collect data about you, use cookies, embed
-              additional third-party tracking, and monitor your interaction with
-              that embedded content, including tracking your interaction with
-              the embedded content if you have an account and are logged in to
-              that website.
-            </p>
-
-            <h2>How long we retain your data</h2>
-            {/*
-          <p>
-            If you leave a comment, the comment and its metadata are retained
-            indefinitely. This is so we can recognize and approve any
-            follow-up comments automatically instead of holding them in a
-            moderation queue.
-          </p> */}
-
-            {/* <p>
-            For users that register on our website (if any), we also store the
-            personal information they provide in their user profile. All users
-            can see, edit, or delete their personal information at any time
-            (except they cannot change their username). Website administrators
-            can also see and edit that information.
-          </p> */}
-            <p>
-              If you use our contact form or email us, the information and its
-              metadata are retained indefinitely. This is done purely for
-              informational and communication purposes.
-            </p>
-
-            <h2>What rights you have over your data</h2>
-
-            {/* <p>
-            If you have an account on this site or have left comments, you can
-            request to receive an exported file of the personal data we hold
-            about you, including any data you have provided to us. You can
-            also request that we erase any personal data we hold about you.
-            This does not include any data we are obliged to keep for
-            administrative, legal, or security purposes.
-          </p> */}
-
-            <p>
-              If you have contacted us, you can request to receive an exported
-              file of the personal data we hold about you, including any data
-              you have provided to us. You can also request that we erase any
-              personal data we hold about you. This does not include any data we
-              are obliged to keep for administrative, legal, or security
-              purposes.
-            </p>
-
-            <h2>Where we send your data</h2>
-
-            {/* <p>
-            Visitor comments may be checked through an automated spam
-            detection service.
-          </p> */}
-
-            <p>We do not send or sell your data to any third parties.</p>
-          </article>
-        </section>
-        <SideBar allPosts={allPosts} />
-      </Container>
-    </>
-  );
-}
-
-export const getStaticProps = async () => {
-  const allPosts = getAllPosts([
-    "title",
-    "date",
-    "slug",
-    "author",
-    "coverImage",
-    "excerpt",
-  ]);
-
-  return {
-    props: { allPosts },
-  };
-};
